+++ src/controllers/PinceController/pince_controller.h
@@ -19,13 +19,13 @@ public:
      * Move the pince vertically
      * 1 = up, 0 = stop, -1 = down
      */
-    void translateVertical(int32_t);
+    void translateVertical(int16_t);
     
     /**
      * Move the pince horizontally
      * 1 = left, 0 = stop, -1 = right
      */
-    void translateHorizontal(int32_t);
+    void translateHorizontal(int16_t);
     
     /**
      * Move the pince horizontally
@@ -78,22 +78,22 @@ public:
     /**
      * Returns the current angle of the pince's poignet
      */
-    int32_t getCurrentPoignetAngle();
+    int16_t getCurrentPoignetAngle();
 
     /**
      * Sets the current angle of the pince's poignet
      */
-    void setCurrentPoignetAngle(int32_t);
+    void setCurrentPoignetAngle(int16_t);
 
     /**
      * Returns the current angle of the pince's flip
      */
-    int32_t getCurrentFlipAngle();
+    int16_t getCurrentFlipAngle();
 
     /**
      * Sets the current angle of the pince's flip
      */
-    void setCurrentFlipAngle(int32_t);
+    void setCurrentFlipAngle(int16_t);
 
     /**
      * Ouvrir la pince
@@ -109,7 +109,7 @@ public:
      * Déplacer le poignet à une position
      * @param angle Angle du poignet
      */
-    void movePoignetPince(int32_t angle);
+    void movePoignetPince(int16_t angle);
 
 private:
     MainController *_mainController = nullptr;
@@ -118,10 +118,10 @@ private:
     Servo _pincePoignetServo;
     Servo _pinceFlipServo;
 
-    int32_t _currentPincePoignetAngle = 0;
-    int32_t _currentPinceFlipAngle = PINCE_FLIP_MIN_ANGLE;
-    int32_t _currentPinceDoigtAngle1 = 0;
-    int32_t _currentPinceDoigtAngle2 = 0;
+    int16_t _currentPincePoignetAngle = 0;
+    int16_t _currentPinceFlipAngle = PINCE_FLIP_MIN_ANGLE;
+    int16_t _currentPinceDoigtAngle1 = 0;
+    int16_t _currentPinceDoigtAngle2 = 0;
     bool _hasReachedReloading = false;
     bool _hasDroppedGamePieces = false;
     bool _areDoigtsOpen = false;
